Update for use with floats now.
git-svn-id: https://svn.aros.org/svn/aros/trunk@3960 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/include/clib/mathffp_protos.h
@@ -19,57 +19,57 @@
 /*
     Prototypes
 */
-AROS_LP1(LONG, SPAbs,
-    AROS_LPA(LONG, fnum1, D0),
+AROS_LP1(float, SPAbs,
+    AROS_LPA(float, fnum1, D0),
     struct MathBase *, MathBase, 9, Mathffp)
 
-AROS_LP2(LONG, SPAdd,
-    AROS_LPA(LONG, fnum1, D1),
-    AROS_LPA(LONG, fnum2, D0),
+AROS_LP2(float, SPAdd,
+    AROS_LPA(float, fnum1, D1),
+    AROS_LPA(float, fnum2, D0),
     struct MathBase *, MathBase, 11, Mathffp)
 
-AROS_LP1(LONG, SPCeil,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, SPCeil,
+    AROS_LPA(float, y, D0),
     struct MathBase *, MathBase, 16, Mathffp)
 
 AROS_LP2(LONG, SPCmp,
-    AROS_LPA(LONG, fnum1, D0),
-    AROS_LPA(LONG, fnum2, D1),
+    AROS_LPA(float, fnum1, D0),
+    AROS_LPA(float, fnum2, D1),
     struct MathBase *, MathBase, 7, Mathffp)
 
-AROS_LP2(LONG, SPDiv,
-    AROS_LPA(LONG, fnum1, D1),
-    AROS_LPA(LONG, fnum2, D0),
+AROS_LP2(float, SPDiv,
+    AROS_LPA(float, fnum1, D1),
+    AROS_LPA(float, fnum2, D0),
     struct MathBase *, MathBase, 14, Mathffp)
 
 AROS_LP1(LONG, SPFix,
-    AROS_LPA(LONG, fnum, D0),
+    AROS_LPA(float, fnum, D0),
     struct MathBase *, MathBase, 5, Mathffp)
 
-AROS_LP1(LONG, SPFloor,
-    AROS_LPA(LONG, y, D0),
+AROS_LP1(float, SPFloor,
+    AROS_LPA(float, y, D0),
     struct MathBase *, MathBase, 15, Mathffp)
 
-AROS_LP1(LONG, SPFlt,
+AROS_LP1(float, SPFlt,
     AROS_LPA(LONG, inum, D0),
     struct MathBase *, MathBase, 6, Mathffp)
 
-AROS_LP2(LONG, SPMul,
-    AROS_LPA(LONG, fnum1, D1),
-    AROS_LPA(LONG, fnum2, D0),
+AROS_LP2(float, SPMul,
+    AROS_LPA(float, fnum1, D1),
+    AROS_LPA(float, fnum2, D0),
     struct MathBase *, MathBase, 13, Mathffp)
 
-AROS_LP1(LONG, SPNeg,
-    AROS_LPA(LONG, fnum1, D0),
+AROS_LP1(float, SPNeg,
+    AROS_LPA(float, fnum1, D0),
     struct MathBase *, MathBase, 10, Mathffp)
 
-AROS_LP2(LONG, SPSub,
-    AROS_LPA(LONG, fnum1, D0),
-    AROS_LPA(LONG, fnum2, D1),
+AROS_LP2(float, SPSub,
+    AROS_LPA(float, fnum1, D0),
+    AROS_LPA(float, fnum2, D1),
     struct MathBase *, MathBase, 12, Mathffp)
 
 AROS_LP1(LONG, SPTst,
-    AROS_LPA(LONG, fnum, D1),
+    AROS_LPA(float, fnum, D1),
     struct MathBase *, MathBase, 8, Mathffp)
